- Add music section - Disable caching for first release (still wip) - Add simple/advanced client settings - Add qr barcode scanning - Disable settings that are currently unused - autoconnect - Many fixes and small additions

+++ trunk/src/com/mediaportal/ampdroid/database/RemoteClientsDatabaseHandler.java
@@ -144,7 +144,7 @@ public class RemoteClientsDatabaseHandler {
    }
 
    public boolean removeRemoteClient(RemoteClient _client) {
-      if (mDatabase != null) {
+      if (mDatabase != null && _client != null) {
          int numDeleted = mDatabase.delete(REMOTE_CLIENTS_TABLE,
                "ClientId=" + _client.getClientId(), null);
          return numDeleted == 1;
